+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/UrlPathHelper.java
@@ -4,7 +4,6 @@ import io.github.poshjosh.ratelimiter.util.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.http.HttpServletRequest;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.util.Objects;
@@ -18,7 +17,7 @@ import java.util.Objects;
  *     <li>HttpServletRequest#getPathInfo() may return null, e.g within Filters.</li>
  *     <li>HttpServletRequest#getRequestURI() is not decoded</li>
  * </ul>
- * <p>For reference, the various path related methods of HttpServletRequest are shown below:</p>
+ * <p>For reference, the various path related methods of RequestInfo are shown below:</p>
  * Application deployed under: /app
  * <br/>Servlet is mapped as: /test%3F/*
  * <br/>URL: http://30thh.loc:8480/app/test%3F/a%3F+b;jsessionid=S%3F+ID?p+1=c+d&p+2=e+f#a
@@ -66,11 +65,11 @@ final class UrlPathHelper {
      * Get the path beginning with the servletPath.
      *
      * If the servletPath is an empty string, this means the servlet used to process this request
-     * was matched using the "/*" pattern. See {@link HttpServletRequest#getServletPath()}.
-     * @param request The HttpServletRequest whose path will be returned
+     * was matched using the "/*" pattern. See {@link RequestInfo#getServletPath()}.
+     * @param request The RequestInfo whose path will be returned
      * @return the path beginning with the servletPath.
      */
-    public String getPathWithinServlet(HttpServletRequest request) {
+    public String getPathWithinServlet(RequestInfo request) {
         final String path = getPathWithinApplication(request);
         final String servletPath = request.getServletPath();
         final String pathWithinServlet;
@@ -93,7 +92,7 @@ final class UrlPathHelper {
         }
         if (LOG.isTraceEnabled()) {
             LOG.trace("Path within servlet: {}, request URI: {}",
-                    pathWithinServlet, request.getRequestURI());
+                    pathWithinServlet, request.getRequestUri());
         }
         return pathWithinServlet;
     }
@@ -104,7 +103,7 @@ final class UrlPathHelper {
      * @param request current HTTP request
      * @return the path within the web application
      */
-    public String getPathWithinApplication(HttpServletRequest request) {
+    public String getPathWithinApplication(RequestInfo request) {
         String contextPath = getContextPath(request);
         String requestUri = getRequestUri(request);
         String path = getRemainingPath(requestUri, contextPath, true);
@@ -120,7 +119,7 @@ final class UrlPathHelper {
     /**
      * Match the given "mapping" to the start of the "requestUri" and if there
      * is a match return the extra part. This method is needed because the
-     * context path and the servlet path returned by the HttpServletRequest are
+     * context path and the servlet path returned by the RequestInfo are
      * stripped of semicolon content unlike the requestUri.
      */
     private String getRemainingPath(String requestUri, String mapping, boolean ignoreCase) {
@@ -183,8 +182,8 @@ final class UrlPathHelper {
      * @param request current HTTP request
      * @return the request URI
      */
-    private String getRequestUri(HttpServletRequest request) {
-        final String uri = request.getRequestURI();
+    private String getRequestUri(RequestInfo request) {
+        final String uri = request.getRequestUri();
         return decodeAndCleanUriString(request, uri);
     }
 
@@ -196,7 +195,7 @@ final class UrlPathHelper {
      * @param request current HTTP request
      * @return the context path
      */
-    private String getContextPath(HttpServletRequest request) {
+    private String getContextPath(RequestInfo request) {
         String contextPath = request.getContextPath();
         if ("/".equals(contextPath)) {
             // Invalid case, but happens for includes on Jetty: silently adapt it.
@@ -208,7 +207,7 @@ final class UrlPathHelper {
     /**
      * Decode the supplied URI string and strips any extraneous portion after a ';'.
      */
-    private String decodeAndCleanUriString(HttpServletRequest request, String uri) {
+    private String decodeAndCleanUriString(RequestInfo request, String uri) {
         uri = removeSemicolonContent(uri);
         uri = decodeRequestString(request, uri);
         uri = getSanitizedPath(uri);
@@ -226,7 +225,7 @@ final class UrlPathHelper {
      * @see URLDecoder#decode(String, String)
      * @see URLDecoder#decode(String)
      */
-    private String decodeRequestString(HttpServletRequest request, String source) {
+    private String decodeRequestString(RequestInfo request, String source) {
         String enc = determineEncoding(request);
         try {
             return URLDecoder.decode(source, enc);
@@ -247,8 +246,8 @@ final class UrlPathHelper {
      * @return the encoding for the request (never {@code null})
      * @see javax.servlet.ServletRequest#getCharacterEncoding()
      */
-    private String determineEncoding(HttpServletRequest request) {
-        String enc = request.getCharacterEncoding();
+    private String determineEncoding(RequestInfo request) {
+        String enc = request.getCharacterEncoding(null);
         if (StringUtils.hasText(enc)) {
             return enc;
         }