+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/WebRateLimiterRegistry.java
@@ -2,7 +2,5 @@ package io.github.poshjosh.ratelimiter.web.core;
 
 import io.github.poshjosh.ratelimiter.RateLimiterRegistry;
 
-import javax.servlet.http.HttpServletRequest;
-
-public interface WebRateLimiterRegistry extends RateLimiterRegistry<HttpServletRequest> {
+public interface WebRateLimiterRegistry extends RateLimiterRegistry<RequestInfo> {
 }